[REFACTOR] dto 컨벤션에 맞게 request와 response로 나누기
