+++ src/main/java/com/devloop/community/request/CommunitySaveRequest.java
@@ -1,21 +1,17 @@
-package com.devloop.scheduleTodo.dto.request;
+package com.devloop.community.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-import java.time.LocalDateTime;
-
 @Getter
 @AllArgsConstructor
-public class ScheduleTodoRequest {
+public class CommunitySaveRequest {
     @NotBlank(message = "제목을 작성해 주세요")
     private String title;
     @NotBlank(message = "내용을 작성해 주세요")
     private String content;
-    @NotNull(message = "시작일을 작성해 주세요")
-    private LocalDateTime startDate;
-    @NotNull(message = "종료일을 작성해 주세요")
-    private LocalDateTime endDate;
+    @NotNull(message = "카테고리를 작성해 주세요")
+    private String category;
 }
