+++ src/main/java/com/devloop/community/request/CommunityUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.devloop.community.dto.request;
+package com.devloop.community.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
@@ -7,11 +7,14 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class CommunitySaveRequest {
+public class CommunityUpdateRequest {
     @NotBlank(message = "제목을 작성해 주세요")
     private String title;
     @NotBlank(message = "내용을 작성해 주세요")
     private String content;
+    @NotNull(message = "해결 상태를 작성해 주세요")
+    private String status;
     @NotNull(message = "카테고리를 작성해 주세요")
     private String category;
 }
+
