+++ src/main/java/com/devloop/community/response/CommunityDetailResponse.java
@@ -1,9 +1,8 @@
-package com.devloop.community.dto.response;
+package com.devloop.community.response;
 
 import lombok.Getter;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
 @Getter
 public class CommunityDetailResponse {
