+++ src/main/java/com/devloop/communitycomment/response/CommentUpdateResponse.java
@@ -1,4 +1,4 @@
-package com.devloop.communitycomment.dto.response;
+package com.devloop.communitycomment.response;
 
 import lombok.Getter;
