+++ src/main/java/com/devloop/communitycomment/service/CommunityCommentService.java
@@ -5,11 +5,11 @@ import com.devloop.common.apipayload.status.ErrorStatus;
 import com.devloop.common.exception.ApiException;
 import com.devloop.community.entity.Community;
 import com.devloop.community.service.CommunityService;
-import com.devloop.communitycomment.dto.CommentResponse;
-import com.devloop.communitycomment.dto.request.CommentSaveRequest;
-import com.devloop.communitycomment.dto.request.CommentUpdateRequest;
-import com.devloop.communitycomment.dto.response.CommentSaveResponse;
-import com.devloop.communitycomment.dto.response.CommentUpdateResponse;
+import com.devloop.communitycomment.response.CommentResponse;
+import com.devloop.communitycomment.request.CommentSaveRequest;
+import com.devloop.communitycomment.request.CommentUpdateRequest;
+import com.devloop.communitycomment.response.CommentSaveResponse;
+import com.devloop.communitycomment.response.CommentUpdateResponse;
 import com.devloop.communitycomment.entity.CommunityComment;
 import com.devloop.communitycomment.repository.CommunityCommentRepository;
 import com.devloop.user.entity.User;
