+++ src/main/java/com/devloop/scheduleBoard/service/ScheduleBoardService.java
@@ -4,10 +4,8 @@ import com.devloop.common.apipayload.status.ErrorStatus;
 import com.devloop.common.exception.ApiException;
 import com.devloop.pwt.entity.ProjectWithTutor;
 import com.devloop.pwt.service.ProjectWithTutorService;
-import com.devloop.scheduleBoard.dto.response.ScheduleBoardResponse;
-import com.devloop.scheduleBoard.entity.BoardAssignment;
+import com.devloop.scheduleBoard.response.ScheduleBoardResponse;
 import com.devloop.scheduleBoard.entity.ScheduleBoard;
-import com.devloop.scheduleBoard.repository.BoardAssignmentRepository;
 import com.devloop.scheduleBoard.repository.ScheduleBoardRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
