+++ src/main/java/com/devloop/scheduleTodo/service/ScheduleTodoService.java
@@ -7,9 +7,9 @@ import com.devloop.pwt.entity.ProjectWithTutor;
 import com.devloop.scheduleBoard.entity.ScheduleBoard;
 import com.devloop.scheduleBoard.repository.BoardAssignmentRepository;
 import com.devloop.scheduleBoard.service.ScheduleBoardService;
-import com.devloop.scheduleTodo.dto.request.ScheduleTodoRequest;
-import com.devloop.scheduleTodo.dto.response.ScheduleTodoResponse;
-import com.devloop.scheduleTodo.dto.response.ScheduleTodoSimpleResponse;
+import com.devloop.scheduleTodo.request.ScheduleTodoRequest;
+import com.devloop.scheduleTodo.response.ScheduleTodoResponse;
+import com.devloop.scheduleTodo.response.ScheduleTodoSimpleResponse;
 import com.devloop.scheduleTodo.entity.ScheduleTodo;
 import com.devloop.scheduleTodo.repository.ScheduleTodoRepository;
 import com.devloop.user.entity.User;
